Crate enroute

🛩️ A parsimonious, framework and IO agnostic HTTP request router.

Enroute is opinionated about being simple with a focus fast and flexible.

Simple

There are very few components to enroute. There is no IO component and there are no framework specific coupling points. It is only composed of types for routing HTTP requests. It leverages the Rust community standard http crate for those types.

Fast

enroute leverage the regex crate's RegexSet feature to filter requests on all patterns in a single pass. This means there is less work for enroute to do for each request. Doing less work wastes less time and less cpu resources.

Flexible

Regexes are universal language for pattern recognition. As such there is incentive to not reinvent a router specific language for http path matching Many routing systems provide a means of extracting information from paths Regex capture groups also do this with built in way of expressing restrictions on the desired type within those capture groups. enroute exposes these captures directly to handlers. Because regular expressions are a universal language it lends it self to more problem domains than can be preemptively designed by an invented syntax.

enroute takes a wholistic view of requests with a notion of a Matcher which is an extension point for matching additional parts of requests. Perhaps on http methods, or header values or query string parameters.

Structs

And

Builder
Collects a list of route definitions

Captures
Captures represents a group of captured strings for a single match.

Or

Router
A compiled set of routes

Traits

Handler
An abstract request handler

Matcher
A secondary means of routing requests via request introspection

Functions

matcher
Utility fn for coercing a capable type into a Matcher

routes
Creates a new Builder