Trait enroute::Matcher

pub trait Matcher<B> {
    fn matches(&self, req: &Request<B>) -> bool;
}

A secondary means of routing requests via request introspection

A handful of Matcher impls are provided for http type values. These also serve as an extension point

Required methods

fn matches(&self, req: &Request<B>) -> bool

Loading content...