Trait enroute::Matcher

pub trait Matcher<B> {
    fn matches(&self, req: &Request<B>) -> bool;

    fn and<M>(self, b: M) -> And<Self, M>
    where
        Self: Sized,
        M: Matcher<B>,
    { ... }

    fn or<M>(self, b: M) -> Or<Self, M>
    where
        Self: Sized,
        M: Matcher<B>,
    { ... }
}

A secondary means of routing requests via request introspection

A handful of Matcher impls are provided for http type values. These also serve as an extension point

Required methods

fn matches(&self, req: &Request<B>) -> bool

Loading content...
Provided methods

fn and<M>(self, b: M) -> And<Self, M> where
Self: Sized,
M: Matcher<B>,

fn or<M>(self, b: M) -> Or<Self, M> where
Self: Sized,
M: Matcher<B>,

Loading content...